REFACTOR: Rename and move `TransformerInterface` to `Transformers\Transformer`
